Added a CMYKtoRGB converter. This should permit some very fine-grained algorithms based on what we expect to be printed.
Unfortunately, CMYK is NOT a default color profile. I would have liked to have relied on the API, but this is too simple a library to deal with all the extra overhead.

The GetBufferedImage call has been updated, the others have not yet followed suit (they will, though). I found a "JSpinner" class, which I've put in the Inkerator for now, but it is currently pseudo-functional.

Will be fixing all this either tonight or tomorrow.


git-svn-id: svn+ssh://tammany.cs.umbc.edu/web/svn/trunk@179 29ff4f17-4be9-44af-a7df-7a8342e46c90