This is a big update, and these are generally frowned upon...sorry. Basically, this is an architectural outline of the scanning and counting processes that happen at the polling site as well as a Scanner Configuration class. It is not finished, but it should give a good idea of how to proceed. Most of the work will be pulling out routines that are already written and dropping them into this new architecture.
To save the scanner config, you create the structure you need in code (based on the options), then then you use XMLEncoder to write it to a file. To read it, you use XMLReader. See the BallotStyleXMLTest.java in the test folder.

The Scanner Config has all the polling site data, and has a Ballot Reader object, which encapsulates all the Geometry, and an Array of Contest objects, which let you figure out how to count and modify them. It is missing some minor config options, e.g. where should you search for the configuration on the disk?

The (absract) BallotReader scans ballots, and has a list of BallotStyles, and an abstract SerialNumberReader which tells you the style to use. It should also (appropriately) do other functions. I think there might need to be an AlignmentMarkReader that encapsulates different types of alignment marks, but for now I left it with the same params we need to do our regular circular alignment marks. The style also tells you which contests the ballots have and where the marks for those contests are located on the ballot. 
 
The BallotReader creates Ballot objects, which contain id, styleid, ballot data, and state (notes and if this ballot should be counted at end of day or later in the process). Based on the style, you can get the contest data on each ballot. 

The counting abstraction relies on the Contest objects. Contests have names of candidates, rules objects for how to interpret marks on the ballots, and tally objects that state the rules for counting and can count based on raw darkmark logic object data. 

To count, you would have a list of ballots, and do something like the following:

foreach contests as c
	Vector<Integer[][]> c_data;
	foreach ballot as b
		if (b.hasContest(c.id))
		{
			c_data.add(b.getContestData());
		}
	endfor
	c.getTallyMethod().tally(c_data);
endfor	

That's not the most efficient, as it's two-pass, but it does work. It might be more efficient to just pass the array of ballot objects and the marking rules to the tally method directly (and that's probably the way to go, we'll see).

git-svn-id: svn+ssh://tammany.cs.umbc.edu/web/svn/trunk@252 29ff4f17-4be9-44af-a7df-7a8342e46c90