Does basic filtering, but still doesn't get the overvotes that can sometimes be counted.
In many IRV systems, the rule is to count an overvote on a rank if all the other candidates have been eliminated when you get to it. Since oSTV doesn't provide any ability to do that we would have to "preclean" the ballots to get the same results. While we do some of the precleaning already to avoid most things, this requires us to essentially count before reporting and will require a little bit more work.

git-svn-id: svn+ssh://tammany.cs.umbc.edu/web/svn/trunk@370 29ff4f17-4be9-44af-a7df-7a8342e46c90