RandomBallotStore is nearly finished.
There is a small bug that I haven't yet been able to fix. It has to do with SAX not liking adding extra characters at the beginning or end. It could be the IOstream class that I'm using adding data, or it could be an off-by-? style of error. I'm not sure.

git-svn-id: svn+ssh://tammany.cs.umbc.edu/web/svn/trunk@375 29ff4f17-4be9-44af-a7df-7a8342e46c90