Added write-in support to the Ballot Object.
When processing images, the write-in rectangle images will be clipped as specified in the BallotStyle object and added to each Ballot object. When processing election results at a later date, Election Judges will use software to read each write-in position,  add each write-In contestant to the Contest, save, and publish the results.

git-svn-id: svn+ssh://tammany.cs.umbc.edu/web/svn/trunk@405 29ff4f17-4be9-44af-a7df-7a8342e46c90