Minor change to the Scanner configuration object, and refactored it to list the getters and setters in order. Updated all the configurator files, too.
The output filenames have all been replaced with output directories---Any writeable directories inside of these locations will get a copy of the data, If there are no writeable directories, we will simply use that directory. The output code for the scanner will use default names for each file (as we decided to standardize last week). 


git-svn-id: svn+ssh://tammany.cs.umbc.edu/web/svn/trunk@464 29ff4f17-4be9-44af-a7df-7a8342e46c90