This is a major revision that is some time in the making.
== (1) Stefan's codebase has been merged into the mainline trunk code. ==

The auditor, authoring, and engine components should now compile and launch with Mavens assembly:assembly command. Not all of the tests work. 

== (2) ballotStandard.jar was reversed and integrated into the mainline trunk code ==

ballotStandard was a jarfile we had been using to read and validate ElectionSpec.xml files. We lost the source code to it before this repository was started and never found it. 

There are several known problems having to do with the validator functions that means this code is likely to be very buggy. Testing will have to resolve all of these issues. We may also need to regenerate ElectionSpec.xsd.

Because there are multiple sets of configuration files, this code will have to be merged with the other config files to create a unified set. I suggest creating an "xml/io/config" module to handle all of the xml file IO.

== (3) All circular references between modules were resolved. ==

Files with references to projects that were not common saw said references moved to common. 

== (4) new module (crypto) ==

You will find confirmation number generator code in here and a test stub for it. 

The intent of this module is to include all crypto-specific or crypto-sensitive code here. Crypto code into the engine will be moved here over time, and the engine will become a GUI front-end.  

This would normally be included in a separate subversion update, but it is distince from and easier to include it with the larger update.

== (5) module "engine" moved to "new-engine" ==

This is an experimental module where we were trying to reimplement the punchscan, eperio, and scantegrity backends into a shared architecture so that they could be used interchangeably with their respective front-ends. Stefan's engine module replaced this module.




git-svn-id: svn+ssh://tammany.cs.umbc.edu/web/svn/trunk@588 29ff4f17-4be9-44af-a7df-7a8342e46c90