ATTN: This code still needs review. See previous commit.
The general algorithm remains the same. This change promotes each byte to an integer, then shifts both ways by 24. Java does not have unsigned numbers, so the promotion/conversion to integer of a negative byte results in something like this:

-128 => 0x80 to 0xFFFFFFFFFFFFFF80

This commit changes it back to 0x80 after the integer conversion. 

git-svn-id: svn+ssh://tammany.cs.umbc.edu/web/svn/trunk@604 29ff4f17-4be9-44af-a7df-7a8342e46c90